:truck: Rename the certificate maker screen module
